[SourceMap] Add unit tests
