+++ package.json
@@ -28,6 +28,7 @@
     "eslint": "^3.12.0",
     "eslint-plugin-babel": "^3.3.0",
     "eslint-plugin-flowtype": "^2.20.0",
+    "eslint-plugin-jest": "^19.0.1",
     "eslint-plugin-mozilla": "0.2.3",
     "eslint-plugin-react": "^6.7.1",
     "husky": "^0.12.0",
@@ -37,7 +38,7 @@
     "stylelint": "^7.4.2"
   },
   "devDependencies": {
-    "lerna": "2.0.0-beta.30",
-    "flow-bin": "^0.37.4"
+    "flow-bin": "^0.37.4",
+    "lerna": "2.0.0-beta.30"
   }
 }
