+++ packages/devtools-source-map/.babelrc
@@ -1,0 +1,6 @@
+{
+  "plugins": [
+    "transform-async-to-generator",
+    "transform-flow-strip-types"
+  ]
+}
