+++ packages/devtools-source-map/package.json
@@ -12,14 +12,31 @@
     "copy-assets": "node bin/copy-assets"
   },
   "dependencies": {
+    "babel-plugin-transform-async-to-generator": "^6.22.0",
+    "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "devtools-client-adapters": "^0.0.21",
     "devtools-config": "^0.0.12",
     "devtools-network-request": "^0.0.14",
+    "jest": "^19.0.2",
     "md5": "^2.2.1",
+    "regenerator-runtime": "^0.10.3",
     "source-map": "^0.5.6",
     "url": "^0.11.0"
   },
   "devDependencies": {
     "devtools-launchpad": "^0.0.56"
+  },
+  "jest": {
+    "rootDir": "src",
+    "testMatch": [
+      "**/tests/**/*.js"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/tests/fixtures/"
+    ],
+    "transformIgnorePatterns": [],
+    "setupFiles": [],
+    "moduleNameMapper": {}
   }
 }
